python3Packages.parameterized: Update for 3.11 support
`parameterized` from pypi does not build for Python 3.11, due to a
legacy dependency on `pytest` which is no longer available and a
dependency on `nose` which is no longer maintained and will not work
for 3.11 onwards. This updates `parameterized` to point at the more
up-to-date version available on GitHub and updates dependencies to
properly import `nose` only for <3.10.

Also updated the license to BSD2 as reflected on Pypi and the repo.